History: use ActiveRecord::Inheritance#base_class
Today, in order to calculate the `history_foreign_key`, `History` walks
up the inheritance chain to try and calculate the base of a (potential)
STI tree. However, it is a default Rails convention to inherit from an
`ApplicationRecord` class. This does not necessarily mean that STI is
being used; the current logic leads to using a foreign key of
`application_record_id` in such a project. Instead of needing to care
about STI, just use `ActiveRecord::Inheritance#base_class`, which does
the right thing in all cases.

Resolve #41

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>